create base solve object and make PicardSolve derive from it and move Picard parameters into PicardSolve #5229
