+++ framework/src/executioners/PicardSolve.C
@@ -16,6 +16,71 @@
 #include "AllLocalDofIndicesThread.h"
 #include "Console.h"
 
+template <>
+InputParameters
+validParams<PicardSolve>()
+{
+  InputParameters params = emptyInputParameters();
+
+  params.addParam<unsigned int>(
+      "picard_max_its",
+      1,
+      "Specifies the maximum number of Picard iterations.  Mainly used when "
+      "wanting to do Picard iterations with MultiApps that are set to "
+      "execute_on timestep_end or timestep_begin. Setting this parameter to 1 turns off the Picard "
+      "iterations.");
+  params.addParam<bool>(
+      "accept_on_max_picard_iteration",
+      false,
+      "True to treat reaching the maximum number of Picard iterations as converged.");
+  params.addParam<bool>("disable_picard_residual_norm_check",
+                        false,
+                        "Disable the Picard residual norm evaluation thus the three parameters "
+                        "picard_rel_tol, picard_abs_tol and picard_force_norms.");
+  params.addParam<Real>("picard_rel_tol",
+                        1e-8,
+                        "The relative nonlinear residual drop to shoot for "
+                        "during Picard iterations.  This check is "
+                        "performed based on the Master app's nonlinear "
+                        "residual.");
+  params.addParam<Real>("picard_abs_tol",
+                        1e-50,
+                        "The absolute nonlinear residual to shoot for "
+                        "during Picard iterations.  This check is "
+                        "performed based on the Master app's nonlinear "
+                        "residual.");
+  params.addParam<bool>(
+      "picard_force_norms",
+      false,
+      "Force the evaluation of both the TIMESTEP_BEGIN and TIMESTEP_END norms regardless of the "
+      "existance of active MultiApps with those execute_on flags, default: false.");
+
+  params.addRangeCheckedParam<Real>("relaxation_factor",
+                                    1.0,
+                                    "relaxation_factor>0 & relaxation_factor<2",
+                                    "Fraction of newly computed value to keep."
+                                    "Set between 0 and 2.");
+  params.addParam<std::vector<std::string>>("relaxed_variables",
+                                            std::vector<std::string>(),
+                                            "List of variables to relax during Picard Iteration");
+
+  params.addParamNamesToGroup("picard_max_its accept_on_max_picard_iteration "
+                              "disable_picard_residual_norm_check picard_rel_tol "
+                              "picard_abs_tol picard_force_norms "
+                              "relaxation_factor relaxed_variables",
+                              "Picard");
+
+  params.addParam<unsigned int>(
+      "max_xfem_update",
+      std::numeric_limits<unsigned int>::max(),
+      "Maximum number of times to update XFEM crack topology in a step due to evolving cracks");
+  params.addParam<bool>("update_xfem_at_timestep_begin",
+                        false,
+                        "Should XFEM update the mesh at the beginning of the timestep");
+
+  return params;
+}
+
 PicardSolve::PicardSolve(Executioner * ex)
   : MooseObject(ex->parameters()),
     PerfGraphInterface(this),
