+++ framework/src/executioners/SolveObject.C
@@ -1,0 +1,27 @@
+//* This file is part of the MOOSE framework
+//* https://www.mooseframework.org
+//*
+//* All rights reserved, see COPYRIGHT for full restrictions
+//* https://github.com/idaholab/moose/blob/master/COPYRIGHT
+//*
+//* Licensed under LGPL 2.1, please see LICENSE for details
+//* https://www.gnu.org/licenses/lgpl-2.1.html
+
+#include "PicardSolve.h"
+
+#include "FEProblem.h"
+#include "Executioner.h"
+#include "MooseMesh.h"
+#include "NonlinearSystem.h"
+#include "AllLocalDofIndicesThread.h"
+#include "Console.h"
+
+SolveObject::SolveObject(Executioner * ex)
+  : MooseObject(ex->parameters()),
+    PerfGraphInterface(this),
+    _executioner(*ex),
+    _problem(*getCheckedPointerParam<FEProblemBase *>(
+        "_fe_problem_base", "This might happen if you don't have a mesh")),
+    _nl(_problem.getNonlinearSystemBase())
+{
+}
